Rewrite 3 (#31)
* fix readme again

* wpm, accuracy

* load history of stats

* Move done channel to viewdata - editor, paragraph

* Inactive flag for stats

* create data on view creation

* add app data module

* wip integrate components

* viewdata for controls

* debug stuff

* move stuff out of main.go

* debug stuff

* choose paragraph new way!

* Go crazy - drop old files

* activate editor

* Setup global logger instance

* add logger.go

* Put back history scroll

* Fix controls

+++ pkg/utils/formula.go
@@ -7,8 +7,8 @@ import (
 
 // CalculateWpm calculates words per minute
 // based on words typed so far and time elapsed
-func CalculateWpm(countw int, secs int) int {
-	return (60 * countw) / secs
+func CalculateWpm(countw int, secs float64) int {
+	return int((60 * float64(countw)) / secs)
 }
 
 // CalculateAccuracy calculates accuracy in a race
